Enable Safari 10 save mangling (#3)
Safari 10 has an issue with `let` and redeclarations; the only way to safely
let uglify mangle and generate a build capable of supporting Safari 10 is to
enable the specific Safari 10 option.

No.

Seriously.

https://github.com/mishoo/UglifyJS2/issues/1753#issuecomment-324814782

There is no noticable bundle size difference with this enabled

| type | safari10: true | safari10: false|
-----------------------------------------
| uncompressed | 494k | 489k |
| brotli  | 130k | 134k |
| gzip | 154k | 148k |
| deflate | 154k | 148k |

So somehow this is making Brotli more efficient. And the only browser that
doesn't support Brotli? Safari 10.

So win-win everywhere!